WIP: starting to add stl_2_volfrac in setup
